LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-used-values-expected.txt


PASS Use the last value from view-timeline-axis if omitted
FAIL Use the last value from view-timeline-inset if omitted assert_equals: expected "0" but got "25"
